add QueryOptions, refactor DistanceStrategy
