--- langchain4j-alloydb-pg/src/main/java/dev/langchain4j/index/VectorIndex.java
@@ -1,11 +1,0 @@
-package dev.langchain4j.index;
-
-import java.util.List;
-
-public interface VectorIndex {
-
-    public String getIndexOptions();
-
-    public List<String> getParameterSettings();
-
-}
