+++ langchain4j-alloydb-pg/src/main/java/dev/langchain4j/store/embedding/index/VectorIndex.java
@@ -1,0 +1,9 @@
+package dev.langchain4j.store.embedding.index;
+
+public interface VectorIndex {
+
+    final String DEFAULT_INDEX_NAME_SUFFIX = "langchainvectorindex";
+
+    public String getIndexOptions();
+
+}
