+++ langchain4j-alloydb-pg/src/main/java/dev/langchain4j/store/embedding/index/query/QueryOptions.java
@@ -1,0 +1,9 @@
+package dev.langchain4j.store.embedding.index.query;
+
+import java.util.List;
+
+public interface  QueryOptions {
+
+    public List<String> getParameterSettings();
+
+}
